Moves film popup mock data into card
